+++ CHANGELOG.md
@@ -7,6 +7,7 @@ WIP:
 - Anonymize password_hash customer attribute in magento 1
 - Add documentation about the `chain` converter
 - Allow disabling converters defined as parameters of other converters
+- Add `setPrefix` and `setSuffix` converters
 
 ## [1.0.0-beta5] - 2019-07-09
 [1.0.0-beta5]: https://git.smile.fr/dirtech/gdpr-dump/compare/1.0.0-beta4...1.0.0-beta5